Refactor chat state management
Refactor chat state management and introduce AssistantRun model

- Introduced new chat state classes: BaseChatState, ClassicChatState, RAGChatState, and AgentChatState to better organize chat state handling.
- Moved AssistantRun model to memory.py for improved structure and clarity.
- Updated imports across multiple files to reflect the new model organization.
- Removed deprecated attributes and streamlined the chat state models for better maintainability.
